requirements.txt
requests
beautifulsoup4 
pandas 
nltk 
faiss-cpu 
sentence-transformers 
ollama 
streamlit 